requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2020.7.22=0
certifi=2020.6.20=py38_0
cloudpickle=1.6.0=py_0
cpuonly=1.0=0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.10.0=py38_0
cytoolz=0.10.1=py38h7b6447c_0
dask-core=2.26.0=py_0
dbus=1.13.16=hb2f20db_0
decorator=4.4.2=py_0
expat=2.2.9=he6710b0_2
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.2=h5ab3b9f_0
glib=2.65.0=h3eb4bd4_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
icu=58.2=he6710b0_3
imageio=2.9.0=py_0
intel-openmp=2020.2=254
jpeg=9b=h024ee3a_2
kiwisolver=1.2.0=py38hfd86e86_0
lame=3.100=h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=he19cac6_1
lz4-c=1.9.2=he6710b0_1
matplotlib=3.3.1=0
matplotlib-base=3.3.1=py38h817c723_0
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.1.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
networkx=2.5=py_0
ninja=1.10.1=py38hfd86e86_0
numpy=1.19.1=py38hbc911f0_0
numpy-base=1.19.1=py38hfa32c7d_0
olefile=0.46=py_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1g=h7b6447c_0
pcre=8.44=he6710b0_0
pillow=7.2.0=py38hb39fc2d_0
pip=20.2.2=py38_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38h05f1152_4
python=3.8.5=h7579374_1
python-dateutil=2.8.1=py_0
pytorch=1.7.1=py3.8_cpu_0
pywavelets=1.1.1=py38h7b6447c_0
pyyaml=5.3.1=py38h7b6447c_1
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
scikit-image=0.16.2=py38h0573a6f_0
scipy=1.5.2=py38h0b6359f_0
setuptools=49.6.0=py38_0
sip=4.19.13=py38he6710b0_0
six=1.15.0=py_0
sqlite=3.33.0=h62c20be_0
tk=8.6.10=hbc83047_0
toolz=0.10.0=py_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cpu
tornado=6.0.4=py38h7b6447c_1
typing_extensions=3.10.0.2=pyh06a4308_0
wheel=0.35.1=py_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
